backend/requirements.txt
fastapi
uvicorn
pymongo[srv]
python-dotenv
openai
llama-index
pydantic
PyPDF2
